refactor: move downloadModels function into modelsManager (#133)
