+++ packages/backend/src/managers/modelsManager.spec.ts
@@ -1,11 +1,14 @@
-import { type MockInstance, beforeEach, expect, test, vi } from 'vitest';
+import { type MockInstance, beforeEach, describe, expect, test, vi } from 'vitest';
 import os from 'os';
 import fs from 'node:fs';
 import path from 'node:path';
+import type { DownloadModelResult } from './modelsManager';
 import { ModelsManager } from './modelsManager';
 import type { Webview } from '@podman-desktop/api';
 import type { CatalogManager } from './catalogManager';
 import type { ModelInfo } from '@shared/src/models/IModelInfo';
+import { RecipeStatusUtils } from '../utils/recipeStatusUtils';
+import type { RecipeStatusRegistry } from '../registries/RecipeStatusRegistry';
 
 const mocks = vi.hoisted(() => {
   return {
@@ -28,8 +31,17 @@ vi.mock('@podman-desktop/api', () => {
   };
 });
 
+let setTaskMock: MockInstance;
+let taskUtils: RecipeStatusUtils;
+let setTaskStateMock: MockInstance;
+
 beforeEach(() => {
   vi.resetAllMocks();
+  taskUtils = new RecipeStatusUtils('recipe', {
+    setStatus: vi.fn(),
+  } as unknown as RecipeStatusRegistry);
+  setTaskMock = vi.spyOn(taskUtils, 'setTask');
+  setTaskStateMock = vi.spyOn(taskUtils, 'setTaskState');
 });
 
 const dirent = [
@@ -303,4 +315,95 @@ test('deleteLocalModel fails to delete the model folder', async () => {
     ],
   });
   expect(mocks.showErrorMessageMock).toHaveBeenCalledOnce();
+});
+
+describe('downloadModel', () => {
+  const manager = new ModelsManager('appdir', {} as Webview, {} as CatalogManager);
+  test('download model if not already on disk', async () => {
+    vi.spyOn(manager, 'isModelOnDisk').mockReturnValue(false);
+    const doDownloadModelWrapperMock = vi
+      .spyOn(manager, 'doDownloadModelWrapper')
+      .mockImplementation((_modelId: string, _url: string, _taskUtil: RecipeStatusUtils, _destFileName?: string) => {
+        return Promise.resolve('');
+      });
+    await manager.downloadModel(
+      {
+        id: 'id',
+        url: 'url',
+        name: 'name',
+      } as ModelInfo,
+      taskUtils,
+    );
+    expect(doDownloadModelWrapperMock).toBeCalledWith('id', 'url', taskUtils);
+    expect(setTaskMock).toHaveBeenLastCalledWith({
+      id: 'id',
+      name: 'Downloading model name',
+      labels: {
+        'model-pulling': 'id',
+      },
+      state: 'loading',
+    });
+  });
+  test('retrieve model path if already on disk', async () => {
+    vi.spyOn(manager, 'isModelOnDisk').mockReturnValue(true);
+    const getLocalModelPathMock = vi.spyOn(manager, 'getLocalModelPath').mockReturnValue('');
+    await manager.downloadModel(
+      {
+        id: 'id',
+        url: 'url',
+        name: 'name',
+      } as ModelInfo,
+      taskUtils,
+    );
+    expect(getLocalModelPathMock).toBeCalledWith('id');
+    expect(setTaskMock).toHaveBeenLastCalledWith({
+      id: 'id',
+      name: 'Model name already present on disk',
+      labels: {
+        'model-pulling': 'id',
+      },
+      state: 'success',
+    });
+  });
+});
+
+describe('doDownloadModelWrapper', () => {
+  const manager = new ModelsManager('appdir', {} as Webview, {} as CatalogManager);
+  test('returning model path if model has been downloaded', async () => {
+    vi.spyOn(manager, 'doDownloadModel').mockImplementation(
+      (
+        _modelId: string,
+        _url: string,
+        _taskUtil: RecipeStatusUtils,
+        callback: (message: DownloadModelResult) => void,
+        _destFileName?: string,
+      ) => {
+        callback({
+          successful: true,
+          path: 'path',
+        });
+      },
+    );
+    setTaskStateMock.mockReturnThis();
+    const result = await manager.doDownloadModelWrapper('id', 'url', taskUtils);
+    expect(result).toBe('path');
+  });
+  test('rejecting with error message if model has NOT been downloaded', async () => {
+    vi.spyOn(manager, 'doDownloadModel').mockImplementation(
+      (
+        _modelId: string,
+        _url: string,
+        _taskUtil: RecipeStatusUtils,
+        callback: (message: DownloadModelResult) => void,
+        _destFileName?: string,
+      ) => {
+        callback({
+          successful: false,
+          error: 'error',
+        });
+      },
+    );
+    setTaskStateMock.mockReturnThis();
+    await expect(manager.doDownloadModelWrapper('id', 'url', taskUtils)).rejects.toThrowError('error');
+  });
 });
